TODO tasks are completed

+++ database/migrations/2022_05_13_220658_create_affiliates_table.php
@@ -18,7 +18,9 @@ return new class extends Migration
             $table->foreignId('user_id');
             $table->foreignId('merchant_id');
             // TODO: Replace me with a brief explanation of why floats aren't the correct data type, and replace with the correct data type.
-            $table->float('commission_rate');
+            // DONE
+            // Since it is financial data, so decimal will work better for accurate outputs when performing deep actions on the value.
+            $table->decimal('commission_rate');
             $table->string('discount_code');
             $table->timestamps();
         });
