+++ database/migrations/2022_05_16_143445_create_orders_table.php
@@ -19,8 +19,10 @@ return new class extends Migration
             $table->foreignId('merchant_id')->constrained();
             $table->foreignId('affiliate_id')->nullable()->constrained();
             // TODO: Replace floats with the correct data types (very similar to affiliates table)
-            $table->float('subtotal');
-            $table->float('commission_owed')->default(0.00);
+            // DONE
+            // Since it is financial data, so decimal will work better for accurate results
+            $table->decimal('subtotal');
+            $table->decimal('commission_owed')->default(0.00);
             $table->string('payout_status')->default(Order::STATUS_UNPAID);
             $table->string('discount_code')->nullable();
             $table->string('external_order_id')->nullable(); //You missed it by mistake, otherwise I am not going to change anything in the database schema
